Use float version of log10 in level meter calculation to avoid overflow issue
Closes #65
Thanks to rlopezdiez

+++ TheAmazingAudioEngine/AEAudioController.m
@@ -1617,8 +1617,8 @@ static BOOL AEAudioControllerHasPendingMainThreadMessages(AEAudioController *THI
         }
     }
     
-    if ( averagePower ) *averagePower = 10.0 * log10((double)group->level_monitor_data.average);
-    if ( peakLevel ) *peakLevel = 10.0 * log10((double)group->level_monitor_data.peak);
+    if ( averagePower ) *averagePower = 10.0 * log10f(group->level_monitor_data.average);
+    if ( peakLevel ) *peakLevel = 10.0 * log10f(group->level_monitor_data.peak);
     
     group->level_monitor_data.reset = YES;
 }
@@ -1632,8 +1632,8 @@ static BOOL AEAudioControllerHasPendingMainThreadMessages(AEAudioController *THI
         _inputLevelMonitorData.monitoringEnabled = YES;
     }
     
-    if ( averagePower ) *averagePower = 10.0 * log10((double)_inputLevelMonitorData.average);
-    if ( peakLevel ) *peakLevel = 10.0 * log10((double)_inputLevelMonitorData.peak);
+    if ( averagePower ) *averagePower = 10.0 * log10f(_inputLevelMonitorData.average);
+    if ( peakLevel ) *peakLevel = 10.0 * log10f(_inputLevelMonitorData.peak);
     
     _inputLevelMonitorData.reset = YES;
 }
